Rename 'orderDetail' to 'order_detail'
