+++ app/controllers/__init__.py
@@ -3,4 +3,4 @@ from .ingredient import IngredientController
 from .order import OrderController
 from .size import SizeController
 from .beverage import BeverageController
-from .orderDetail import OrderDetailController
+from .order_detail import OrderDetailController
